+++ src/backend.rs
@@ -12,13 +12,20 @@ use crate::spec::NUM_REGISTERS;
 use core::arch::asm;
 use core::fmt::Debug;
 use core::num::NonZeroU8;
-use core::ptr::{self, read_volatile, write_volatile};
+use core::ptr::NonNull;
 
 mod private {
     pub trait Sealed {}
 }
 
 /// Abstraction over register addresses in [`Backend`].
+///
+/// # Safety
+///
+/// All implementations and instances of this trait are created within this
+/// crate and do follow all safety invariants. API users don't get access to the
+/// underlying register addresses, nor can they construct one themselves, as this
+/// type et al. are sealed.
 pub trait RegisterAddress: Copy + Clone + Debug + Sized + private::Sealed {
     /// Adds a byte offset onto the base register address.
     fn add_offset(self, offset: u8) -> Self;
@@ -49,7 +56,7 @@ impl private::Sealed for PortIoAddress {}
 ///
 /// See [`RegisterAddress`].
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Hash)]
-pub struct MmioAddress(pub(crate) *mut u8);
+pub struct MmioAddress(pub(crate) NonNull<u8>);
 
 // SAFETY: `Uart16550` is not `Sync`, so concurrent access from multiple
 // threads is not possible through this type's API alone. Implementing `Send`
@@ -239,6 +246,67 @@ impl Backend for PioBackend {
     }
 }
 
+/// Arch-specific quirks to access hardware.
+///
+/// On MMIO-access on aarch64, LLVM may emit instructions that are not properly
+/// virtualizable. We therefore need to be more explicit about the instruction.
+/// More info: <https://github.com/rust-lang/rust/issues/131894>
+mod arch {
+    use crate::backend::MmioAddress;
+    #[cfg(any(doc, not(target_arch = "aarch64")))]
+    use core::ptr;
+
+    /// Wrapper around [`ptr::read_volatile`].
+    #[cfg(target_arch = "aarch64")]
+    #[inline(always)]
+    pub unsafe fn mmio_read_register(address: MmioAddress) -> u8 {
+        let ptr = address.0.as_ptr();
+        let ret: u8;
+        // SAFETY: Caller ensures the address is valid MMIO memory.
+        unsafe {
+            core::arch::asm!(
+                "ldrb {ret:w}, [{ptr}]",
+                ptr = in(reg) ptr,
+                ret = out(reg) ret,
+                options(nostack, preserves_flags)
+            );
+        }
+        ret
+    }
+
+    /// Wrapper around [`ptr::read_volatile`].
+    #[cfg(not(target_arch = "aarch64"))]
+    #[inline(always)]
+    pub unsafe fn mmio_read_register(address: MmioAddress) -> u8 {
+        // SAFETY: Caller ensures the address is valid MMIO memory.
+        unsafe { ptr::read_volatile(address.0.as_ptr()) }
+    }
+
+    /// Wrapper around [`ptr::write_volatile`].
+    #[cfg(target_arch = "aarch64")]
+    #[inline(always)]
+    pub unsafe fn mmio_write_register(address: MmioAddress, value: u8) {
+        let ptr = address.0.as_ptr();
+        // SAFETY: Caller ensures the address is valid MMIO memory.
+        unsafe {
+            core::arch::asm!(
+                "strb {val:w}, [{ptr}]",
+                val = in(reg) value,
+                ptr = in(reg) ptr,
+                options(nostack, preserves_flags)
+            );
+        }
+    }
+
+    /// Wrapper around [`ptr::write_volatile`].
+    #[cfg(not(target_arch = "aarch64"))]
+    #[inline(always)]
+    pub unsafe fn mmio_write_register(address: MmioAddress, value: u8) {
+        // SAFETY: Caller ensures the address is valid MMIO memory.
+        unsafe { ptr::write_volatile(address.0.as_ptr(), value) }
+    }
+}
+
 /// MMIO-mapped UART 16550.
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Hash)]
 pub struct MmioBackend {
@@ -264,23 +332,23 @@ impl Backend for MmioBackend {
 
     #[inline(always)]
     unsafe fn _read_register(&mut self, address: MmioAddress) -> u8 {
-        debug_assert_ne!(address.0, ptr::null_mut());
         debug_assert!(address >= self.base());
         let upper_bound_incl = (NUM_REGISTERS - 1) * usize::from(u8::from(self.stride));
-        debug_assert!(address.0 <= self.base().0.wrapping_add(upper_bound_incl));
+        // Address is in the device's address range
+        debug_assert!(address.0.as_ptr() <= self.base().0.as_ptr().wrapping_add(upper_bound_incl));
 
         // SAFETY: The caller ensured that the MMIO address is safe to use.
-        unsafe { read_volatile(address.0) }
+        unsafe { arch::mmio_read_register(address) }
     }
 
     #[inline(always)]
     unsafe fn _write_register(&mut self, address: MmioAddress, value: u8) {
-        debug_assert_ne!(address.0, ptr::null_mut());
         debug_assert!(address >= self.base());
         let upper_bound_incl = (NUM_REGISTERS - 1) * usize::from(u8::from(self.stride));
-        debug_assert!(address.0 <= self.base().0.wrapping_add(upper_bound_incl));
+        // Address is in the device's address range
+        debug_assert!(address.0.as_ptr() <= self.base().0.as_ptr().wrapping_add(upper_bound_incl));
 
         // SAFETY: The caller ensured that the MMIO address is safe to use.
-        unsafe { write_volatile(address.0, value) }
+        unsafe { arch::mmio_write_register(address, value) }
     }
 }